refactor: 전체 렌더 함수 분리
